+++ tests/tracing/utils/test_timeout.py
@@ -7,7 +7,6 @@ import pytest
 import mlflow
 from mlflow.entities.span_event import SpanEvent
 from mlflow.entities.span_status import SpanStatusCode
-from mlflow.pyfunc.context import Context, set_prediction_context
 from mlflow.tracing.export.inference_table import _TRACE_BUFFER, pop_trace
 from mlflow.tracing.trace_manager import _Trace
 from mlflow.tracing.utils.timeout import MlflowTraceTimeoutCache
@@ -100,6 +99,8 @@ def test_trace_halted_after_timeout(monkeypatch):
 def test_trace_halted_after_timeout_in_model_serving(
     monkeypatch, mock_databricks_serving_with_tracing_env
 ):
+    from mlflow.pyfunc.context import Context, set_prediction_context
+
     monkeypatch.setenv("MLFLOW_TRACE_TIMEOUT_SECONDS", "3")
 
     # Simulate model serving env where multiple requests are processed concurrently
